api/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosmtplib==3.0.2",
    "alembic==1.16.5",
    "amqp==5.3.1",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asgiref==3.9.2",
    "asyncpg==0.30.0",
    "bcrypt==4.0.1",
    "billiard==4.2.2",
    "blinker==1.9.0",
    "celery==5.5.3",
    "certifi==2025.8.3",
    "click==8.2.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "colorama==0.4.6",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.11",
    "fastapi-cloud-cli==0.1.5",
    "fastapi-mail==1.5.0",
    "flower==2.0.1",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "humanize==4.13.0",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "kombu==5.5.4",
    "loguru==0.7.3",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "packaging==25.0",
    "passlib==1.7.4",
    "prometheus-client==0.23.1",
    "prompt-toolkit==3.0.52",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "redis==6.4.0",
    "rich==14.1.0",
    "rich-toolkit==0.15.1",
    "rignore==0.6.4",
    "sentry-sdk==2.37.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "sqlmodel==0.0.24",
    "starlette==0.47.3",
    "tornado==6.5.2",
    "typer==0.17.4",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "vine==5.1.0",
    "watchfiles==1.1.0",
    "wcwidth==0.2.14",
    "websockets==15.0.1",
    "win32-setctime==1.2.0",
]
